Create README.md
In this project, we worked with the Teach For America admissions dataset to predict if an applicant will complete the admissions process. The target variable is Completed Admissions Process, coded as 0 and 1. A value of 1 means the applicant followed through with the full process. A value of 0 means they did not complete it.
The main goal of our analysis was to use the methods we learned in class to build several predictive models and compare how well they perform. We focused on:
	kNN (k nearest neighbors)
	Naive Bayes
	Decision Tree with C5 point zero  
	SVM (Support Vector Machine) with radial kernel  
	ANN (Artificial Neural Network)